up : доработаны сервисы (все ещё требуют доработки), проведена работа с версиями, обновлены мапперы и прочие мелкие правки

+++ src/main/java/ru/skypro/homework/dto/CommentDTO.java
@@ -8,7 +8,7 @@ public class CommentDTO {
     private int author;
     private String authorImage;
     private String authorFirstName;
-    private int createdAt;
+    private Long createdAt;
     private int pk;
     private String text;
 }
